requirements.txt
pdf2image
Pillow
pytesseract
Flask
torch
transformers
